requirements.txt
kafka-python==2.2.15
requests==2.31.0
python-dotenv
apache-airflow-providers-apache-spark==4.6.0